Input: keyreset: panic if platform need to
Add extra helper to panic the whole kernel by keyboard combination
Add dump system status in keyreset In android's keyreset feature,
it only supports restarting system when combination keys are pressed,
but some times we need to dump system status instead of restarting
system.
So we improve keyreset feature to support this requirement.
the suggested implementation of dump function is:
1) call show_state() and dump_stack() when combination keys
are pressed in 5~10 seconds;
2) call panic() when combination keys are pressed in 10+ seconds;

Change-Id: I4b7c602acd5be44c6c0cad8e0a5ea4d010a89505
Signed-off-by: LuCao <lucao@marvell.com>
Signed-off-by: Lei Wen <leiwen@marvell.com>
Signed-off-by: Xiaoguang Chen <chenxg@marvell.com>