usb: gadget: f_mass_storage: hide the caching mode page to improve write speed on Win 7
The slower write speed form Win 7 to UMS is because in some case Win 7
is sending WRITE COMMAND with FUA = 1, when UMS driver receives the command,
it will set the file flag with O_SYNC. With the O_SYNC flag, the vfs_write()
will disable the cache and wait until the file block is really written into
the file system. So under this situation, the write speed is very slow.

Linux PC does not handle the caching mode page, it will not send WRITE COMMAND
with FUA = 1. The  write speed from Linux PC to the UMS device is as fast as
to the card reader + TF card.

The best way to solve this problem is to hide the caching mode page of the
UMS. If the Windows 7 does not find a write cache, it will send WRITE COMMAND
with FUA = 0, and this will allow the UMS to handle the cache affairs itself.

Change-Id: I6f01e947abfa5bc3e1047b6e8dc3d4656ee84492
Signed-off-by: Leo Song <liangs@marvell.com>
Signed-off-by: Yu Xu <yuxu@marvell.com>
Signed-off-by: Carl Peng <carlpeng@marvell.com>