mmc: sdhci: use CONFIG_SDHCI_USE_LEDS_CLASS instead
Orignally, the Macro of "SDHCI_USE_LEDS_CLASS" is defined and used within
sdhci.c, and it is not flexible to be disabled. Sometimes we still don't
want to enable this feature even LEDS_CLASS is supported.

Now we use "CONFIG_SDHCI_USE_LEDS_CLASS" instead, and it can be handled
by "Kconfig".

Change-Id: I90af66da0aa8209c081401dba31a8b5f312ec111
Signed-off-by: Jialing Fu <jlfu@marvell.com>
Signed-off-by: Xiang Wang <wangx@marvell.com>