mmc: core: set timing mode as enum type
The old version define the timing type as
a normal decimal value. And if the mmc/sd
spec define the new type, the code should
add more definitions and assign a new value.

This patch set the timing mode as enmu type.
In this type, we could add new mode without
caring its value.

Change-Id: I215a1150a24edc14e67fe9a56fe75ac63a5cc19f
Signed-off-by: Tim Wang <wangtt@marvell.com>