mmc: block: retry writing if Data CRC error
Currently code base, it would try to single block read if bus Read CRC
error, but not error handler for writing.

But it is not good to try to single block write, as the performance
would be downgrade so much. This patch, it try to re-send the write
command if CRC error.

Verified in pxa988/986 FF phones, the patch work well indeed to avoid
issues that eMMC disk is re-mounted as read-only because data crc error
during runtime.

Change-Id: Id4c4034f9eb2671a850676a4b6ba08b76beedea6
Signed-off-by: Jialing Fu <jlfu@marvell.com>