ARM: mmp: add ARM ERRATA 794322 fix for 1088/1L88
ARM ERRATA 794322 ask all r0p2/r0p3 Cortex-A7 MPCore processor
with an L2 cache to apply the following WR:
Inserting both of the following after the SCTLR.C bit is cleared
to 0, and before the caches are cleaned or invalidated:
1) A TLBIMVA operation to any address.
2) A DSB instruction.

On 1088/1L88, we need clear SCTLR.C bit before C2/D2 entry.

Quote the reply from ARM support for the address's validation:
Any address will do, it does not have to be a valid address.

Change-Id: I312fa19c8c7487036531df9428e9e27eae03d18e
Signed-off-by: Yilu Mao <ylmao@marvell.com>
Signed-off-by: Tim Wang <wangtt@marvell.com>