mmc: sdhci-pxav3: change 74 clock wait condition
1. Current code have a problem that won't wait
the SDCE_MISC_INT as the break condition.
This patch fix this logic issue.

2. Fix the warning when SD/SDIO card is not inserted:
sdhci-pxav3 sdhci-pxav3.1: 74 clock interrupt not cleared
Rootcause:
We throught the lowest freq is 100KHz, so we set the maxinum
delay time to 74*10us. But when we try to set 100KHz, in fact,
the actual clock is less than 100KHz after devided. So
orignally delay value is too small.
74*20us is safe now.

Change-Id: I38ccd54218d7d5dfa5d8c6a94249381191c6092a
Signed-off-by: Tim Wang <wangtt@marvell.com>
Signed-off-by: Jialing Fu <jlfu@marvell.com>
Signed-off-by: Xiang Wang <wangx@marvell.com>