of: fix CONFIG_CMDLINE_EXTEND
strlcat takes the size of the buffer, not the number of characters
to concatenate.  If the size of the device tree command line p is
larger than the CONFIG_CMDLINE string data, then strcat(data, p, l)
will hit a BUG_ON because strlen(data) > l.

Replace the second strlcat with a strncpy plus a manual null
termination.

Also rearrange the code to reduce indent depth to make it more
readable, and replace data with a char *cmdline to avoid extra
casts.

Change-Id: Icef33c5e2a9617a3b0f824ad7d23dd8203c982aa
Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: Qiming Wu <wuqm@marvell.com>