TTY: serial: fix termios lost in uart_resume_port
If uart port is a console, when it's opened as a tty device,
console.cflag will be copyed to tty_struct->termios and itself
will be set to 0. tty_struct->termios will be used afterwards.
But after this tty is closed, termios info will be lost too.
The cflag of the console will be set to 0 in uart_port_resume, not
our orignal value. This is not what we want.
The solution is to keep console.cflag and use it in resume if
tty_struct->termios is not available.

Change-Id: Ia081e2336f11863d726ab6fc864a06f80e438fbb
Signed-off-by: Xiang Wang <wangx@marvell.com>
Signed-off-by: Wenbing Wang <wangwb@marvell.com>