usb: gadget: use PHY driver for udc
Originaly, udc driver will call the callbacks in platform data
for PHY initialization and shut down.
With PHY driver, it will call the APIs provided by PHY driver
for PHY initialization and shut down. It removes the callbacks
in platform data, and at same time it removes one block in the
way of enabling device tree for udc driver.

Change-Id: Ia1aa7e33d1cc4d9cd17ce6ffe10058ab0fb43475
Signed-off-by: Chao Xie <chao.xie@marvell.com>