usb: ehci: ehci-mv: use PHY driver for ehci
Originaly, ehci driver will call the callbacks in platform data
for PHY initialization and shut down.
With PHY driver, it will call the APIs provided by PHY driver
for PHY initialization and shutdown. It removes the callbacks
in platform data, and at same time it removes one block in the
way of enabling device tree for ehci driver.

Change-Id: I4fb2ad015cc8c5a6ea62196d3a191373a3895fa1
Signed-off-by: Chao Xie <chao.xie@marvell.com>
Signed-off-by: Dongjiu Geng <djgeng@marvell.com>