usb: phy: use USB2 PHY driver for otg
Originaly, otg driver will call the callbacks in platform data
for PHY initialization and shut down.
With PHY driver, it will call the APIs provided by PHY driver
for PHY initialization and shutdown. It removes the callbacks
in platform data, and at same time it removes one block in the
way of enabling device tree for otg driver.

Change-Id: I759117d9c7d836babdd780cd2e13ea2572243260
Signed-off-by: Chao Xie <chao.xie@marvell.com>