mm: Don't put CMA pages on per-cpu lists
Currently, when freeing 0 order pages, CMA pages are treated
the same as regular movable pages, which means they end up
on the per-cpu page list. This means that the CMA pages are
likely to be allocated for something other than contigous
memory. This increases the chance that the next alloc_contig_range
will fail because pages can't be migrated.

Given the size of the CMA region is typically limited, it is best to
optimize for success of alloc_contig_range as much as possible.
Do this by freeing CMA pages directly instead of putting them
on the per-cpu page lists.

This patch is from Linaro.

Change-Id: Ie66e4ee4a3c8a2aaa3dcd6dd7d96d946eed22808
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>
Signed-off-by: Bin Wang <binw@marvell.com>
Signed-off-by: Lisa Du <cldu@marvell.com>