Makefile: remove the git local version
Remove the git local version to avoid mismatch with modules.
Make sure the kernel version is 3.10.0 instead of the
dynamically auto-generated one such as 3.10.0-245965-g114fc31

Port log by Yilu:
This patch is not same as Xiaofan's patch on kernel 3.10 because
this part of code has modified by Open Source during 3.10 -> 3.14.
But the purpose of this patch is same as Xiaofan's.

Change-Id: I1a0cbc1c81e9b54ecc3b62093e005efa45ea2b98
Signed-off-by: Xiaofan Tian <tianxf@marvell.com>
Signed-off-by: Yilu Mao <ylmao@marvell.com>