mmc: sdhci-pxav3: add dma_mask
add dma_mask to MASK_BIT 32:
this value is used to check whether this physical address is capable
for DMA usage.
for sdhci-v3, dma addr value in both register(SDMA) and descriptor
(ADMA) are only 32 bit, so address lower than 4G is not capable for
DMA. So we set it to 32 here.
For further version of sd host controller, if DMA addr could be set
in higher number, we could set it to higher value.

Change-Id: I9eb0c2b84fb3be5984d94642094c71f461aab848
Signed-off-by: Zhou Zhu <zzhu3@marvell.com>
Signed-off-by: Xiang Wang <wangx@marvell.com>