dmabuf: add meta data attachment
we found there'd be varied specific data connected to a dmabuf, like
the iommu mapping of buffer, the dma descriptors (that can be shared
between several components). Though these info might be able to be
generated every time before dma operations, for performance sake,
it's better to be kept before really invalid.

Change-Id: I89d43dc3fe1ee3da91c42074da5df71b968e6d3c
Signed-off-by: Bin Wang <binw@marvell.com>