arm64: mm: setup dma coherent static memory mapping
The memory space from PAGE_OFFSET to 0xffffffffffffffff was all for kernel
logical memory map, 256GB for 4KB page and 2TB for 64KB page, which is far
more larger than available ddr ram size nowadays.

This patch cuts this space in half, and setup dma coherent mapping of the
memory in the high half space. That is for 4KB page:
0xffffffc000000000~0xffffffdfffffffff : Normal kernel logical memory map
0xffffffe000000000~0xffffffffffffffff : DMA coherent memory map

Change-Id: Iae5695d44ff2b2c9525b87f0f2293677135e2372
Signed-off-by: Bin Wang <binw@marvell.com>
Signed-off-by: Lisa Du <cldu@marvell.com>