mmc: sdhci: allow driver hold wake_lock after CMD
We meet a throughput issue in Iperf case:
Currently, the user APP iperf doesn't claim system suspend lock, so the
system may enter suspend state even iperf is running if nobody claim
the wake lock manually. And the iperf throughput data would decrease
because of system suspend.

Although it is the responsibility of APP to claim wake lock to prevent
suspend, here still provides a solution:
Enabling timeout wakelock after each CMD: if R/W are called frequently,
wake lock is held without released, so system has no chance to enter
suspend. But if the R/W stops for a while, the wakelock will be released
and system has chance to enter suspend again.

And this patch provide a quirks2 flag to enable/disable this feature.

Change-Id: I826a4ef1f26d940e0b14fcd41693159b221ed10e
Signed-off-by: Jialing Fu <jlfu@marvell.com>
Signed-off-by: Xiang Wang <wangx@marvell.com>