iommu/arm-smmu: allow 42-bit stage-1 output size with 64K pages
The output size of stage-1 is currently limited by the input size of
stage-2, which is further limited by VA_BITS since we make use of the
standard pgd_alloc functions for creating page tables.

This patch ensures that we use VA_BITS instead of a hardcoded '39'
for the stage-1 output size limit.

Change-Id: Ideeb4727fb8e21531733f2e6d3a3e6951afd99dd
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Lisa Du <cldu@marvell.com>