iommu/arm-smmu: caps: add IOMMU_CAP_INTR_REMAP capability
MSIs are just seen as bog standard memory writes by the ARM SMMU, so
they can be translated (and isolated) in the same way.

This patch adds the IOMMU_CAP_INTR_REMAP capability to the ARM SMMU
driver and reworks our capabaility code so that we don't assume the
caps are organised as bits in a bitmask (since this isn't the intention).

Change-Id: Ie5bc2558eb22e0b03f1e7df84af55814e3e8c305
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Lisa Du <cldu@marvell.com>