iommu/arm-smmu: fix capability checking prior to device attach
If somebody attempts to check the capability of an IOMMU domain prior to
device attach, then we'll try to dereference a NULL SMMU pointer through
the SMMU domain (since we can't determine the actual SMMU instance until
we have a device attached).

This patch fixes the capability check so that non-global features are
reported as being absent when no device is attached to the domain.

Change-Id: I011208c0e7f4c360ac13e7eee562d8b127f2eba1
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Lisa Du <cldu@marvell.com>