media: SoCISP: vdev enable buffer-driven stream_on
For many ISP, H/W can be changed to a temporary working state that
after the job is done, it can automatically change back to idle
state. This behavior can NOT be tracked by V4L2 API because V4L2
stream state is a S/W concept.
To allow ISP driver support this kind of H/W behavior, a buffer-
driven stream state is added to SoCISP-vdev, so that H/W stream
state can be tracked.

Change-Id: I25374f591950b3b0f9c88962f480746315e2309f
Signed-off-by: Jiaquan Su <jqsu@marvell.com>