PM / devfreq: add PM Qos support
Even if the performance of a device is controlled properly with devfreq,
we still need to get PM-QoS inputs in order to meet the required
performance or special limitations sometimes.

In order to handle QoS requests in devfreq devices, the devfreq device
drivers only need to prepare the freq-table with QoS class along with
devfreq_add_device() call.

Other device drivers that have QoS request can use the freq value
directly to require min/max QoS.

Change-Id: I03bf196c99c19484f58f86fb281972e2e0f7adc8
Signed-off-by: Qiming Wu <wuqm@marvell.com>