PM / devfreq: set deferred workqueue timer to cpu 0 only
This patch is to fix for the following abnormal behavior:
The DDR profiler may only be triggerred after more than 10s.

The reason for such issue is that the profiler, using devfreq mechanism,
is using deferred work.  In devfreq code, the timer is not set to
specific cpu.  Whenthe timer is set to cpu other than cpu 0. The next
timer on that cpu may only happen after long time.

This patch changes open source code to make sure the deferred timer is
only set to cpu 0.  Since all of our interrupts are routed to cpu 0, the
deferred timer always has chance to be handled in a reasonable interval.
And this behavior is aligned with cpufreq, whose timer is also set to cpu 0.

By observation, the longest interval now is 6s when system is idle.

Change-Id: I57a5532cdaaf5e759b34a044dd4efc591f244779
Signed-off-by: Qiming Wu <wuqm@marvell.com>