arm64: mcpm: fix cpu id in mcpm_plat_cpu_power_up
mcpm_plat_cpu_power_up() passed cpu_logical_map(cpu)
to the low level function, that needs the
HWID (MPIDR[15:0]).
cpu_logical_map(cpu) can be used when cpu is the logical
CPU number, but in this function cpu is the core number
in the cluster (which is also given).
This works only when cluster=0.
The correct operation is to reconstruct HWID from
cpu/cluster.

NOTE, that the new .release method has the same issue.

Change-Id: I1ee89a1a1f0d98bd8a92bd55d40d8a4a90eecc51
Signed-off-by: Anton Eidelman <antone@marvell.com>