usb: gadget: fix usb controller reset repeatly.
The USB controller will reset repeatly while enumerated.
It exceeds 3ms when the device driver run to the point of
reading the portsc register. According to the USB spec.
This case is recognized as reset fail. So the driver
reset the USB contoller again. After debug, found that
the cause is that print operation(pr_info) in function
android_work. The pr_info occupied the CPU for a long time
resulting exceeding 3ms. So changed the pr_info to pr_debug
to save time.

Change-Id: I63c510b5ae85c1477095374727ad00e86e899202
Signed-off-by: Dongjiu Geng <djgeng@marvell.com>
Signed-off-by: Chao Xie <chao.xie@marvell.com>