drivers: cpuidle: CPU idle ARM64 driver
This patch implements a simple CPU idle driver for ARM64 machines, used for
demonstration purposes.

C1 state is a simple standbywfi, whereas C2 and C3 state hinges on the suspend
back-end protocol call to enter/exit the state which implies context save and
restore operations.

C2/C3 states are power down states, and the C-state index is used to
communicate to the CPU ops suspend method the parameter defining the
state to be entered.

Cache flushing is implemented in the CPU operations suspend back-end and the
driver code relies on cpu_suspend API to save/restore core registers that are
reset when the CPU hits C2/C3 states.

Change-Id: I044f8d736c6c428fc7a3dd01982947e88f66c76a
Signed-off-by: Fangsuo Wu <fswu@marvell.com>