arm64: mcpm: refine mcpm coupled state checking
mcpm's last man needs to know two values according to the coupled
state, one thing is weather the cluster need to be powered off or not,
another thing is the h/w power state need to set for power controller.

Check weather the cluster is powered off or not, we need make sure
all cores are powered off and the coupled state is bigger than
cpu_powerdown state.

If the cluster is powered off, the cpu need calculate D state with
constraint. If the cluster is powered on, the cpu need calculater C
state, so it cannot bigger than cpu_powerdown state.

Change-Id: Ib3b7d77e577fe48ce99028de91accf7e30f59b0d
Signed-off-by: Fangsuo Wu <fswu@marvell.com>