kthread: disable preempt in __kthread_parkme
When we call complete, it will wakeup the hotplug thread. Then there
will be a chance to do preempt. So it is easy to occur such case if
the hotplug thread is scheduled to run on this hotplug-out cpu after
I increase the hotplug down thread priority.

CPU0				CPU2
cpu_down()
 kthread_park()
  set_bit(KTHREAD_SHOULD_PARK, &kthread->flags)
  wake_up_process(k)
				__kthread_parkme()
				 test_and_set_bit(KTHREAD_IS_PARKED, &self->flags)
				 complete(&self->parked); <= will wake up hotplug thread
				 (chagne to run hotplug thread, and move to CPU0)

   cpu_up()
    __kthread_bind() <= polling until tsk->on_rq == 0

     schedule()
      prev->on_rq = 0 <= polling successfully

This patch disables preempt temporally when we call complete, then when
parkme call schedule function explicitly, this thread will be removed
from run queue.

Change-Id: Ie9529fa41936d5e89301af40c5b317edb3055f9d
Signed-off-by: Xiaofan Tian <tianxf@marvell.com>
Signed-off-by: Jane Li <jiel@marvell.com>