fs: pstore: ram: flush cache to prevent data corruption
Since it is possible that we have dirty data in the cacheline before
we remap ramoops buffer as noncacheable. So that when the previous
dirtied cache line being swap into ddr eventually, it may corrupt
the current ramoops data which kernel is operating over. It is for
kernel's ramoops data is directly through the cpu's cache, so that
it cannot affect/aware cache's status.

Invalidate this range to prevent the potential data corruption.

Change-Id: Id5d49fc8a485173eba52410c4800a1c945f1b48e
Signed-off-by: Lei Wen <leiwen@marvell.com>