video: mmp: refine disp onoff function
There are several scenarioes about onoff sequence, refine them as below:

1. Fix order of onoff sequence in overlay_set_onoff:
When Android is going to power off mode, this onoff function is
called and the off sequence in done correctly - DMA off, path off.
When power-on, the onoff function is called with on, and the same
sequence is done: DMA on, path on.
However, the DMA on is failing since the display clock is turned off,
and bad values are read/write to registers. The display clock is
enabled only in path on function and therefore the sequence
should be changed to path on -> DMA on

2. use set_status to handle dynamic onoff, and reduced_on which means kernel need to
keep some uboot display setting and do a reduced lcd on sequence.

3. blank idle signal will block DDR frequency change in suspend, so disable
it when enter suspend and enable it when resume, which is done in
path_enabledisable function.

4. In mmpfb_blank(), when blank LCD by user, need update the LCD status as suspend/
resume,too, avoid user get wrong status for LCD module.

Change-Id: I570ca15ef583e7aea7bb8ffe1c9adcc8dc06c4ec
Signed-off-by: Qing Zhu <qzhu@marvell.com>