video: mmp: register pm runtime ops for disp
Register pm runtime ops for ctrl and fb, after enable suspend&resume,
there are some scenarios also need to be concerned as below:

1. add protection for registers access, in some scenarios, user case still
access registers unexpected when suspend, which would cause system hang. So
add protection for that.

2. add store and recovery support for disp regs, For some platform (with DC4),
the retention feature was removed, so we need store and recovery registers for LPM.

3. add function which skip power on in kernel, if uboot enable display
and send cmd "fbmem..", kernel will follow uboot's display details; or
it will do a overall power on.

Change-Id: I0b562f774b6ca0fa2e1aecce18bc3de4d3f34ef1
Signed-off-by: Qing Zhu <qzhu@marvell.com>