clk: adjust clk enable point for clks with flag CLK_SET_RATE_ENABLED
This adjust is major for clock notifier dvfs usage.
For clock with flag CLK_SET_RATE_ENABLED, when it is disabled, and
set_rate is called, it requests enable the clock at first, then set_rate,
disable it at last.
But previously enable point is after PRE_RATE_CHANGE notifier, which
will lead to clock is enable at old rate(voltage voting), but scaling
to higher frequency when freq-chg.
So now adjust the sequence to enable clock at first, then set notifier
to make sure voltage voting is expected when clock is enable and scaling
to new rate.

Change-Id: I6b96d6fae1140d295d5e0c8fee1ce750003cfe2f
Signed-off-by: Zhoujie Wu <zjwu@marvell.com>