dma: mmp_pdma: add some chan paramaters initialization
Members of "mmp_pdma_chan" are all set to 0 when initialization in
mmp_pdma_chan_init(). That's OK if a channel is requested for the
first time. But if a channel is requested then released, some
paramaters are not set to init values. It will cause unexpected
issue when this channel is requested afterwards.
Sometimes dma request number is not needed (e.g. memory copy), our
driver may set a wrong DMA Mapping register in this case.

Change-Id: I8181013df8000eb679ca6648673c871c4e319626
Signed-off-by: Xiang Wang <wangx@marvell.com>
Signed-off-by: Wenbing Wang <wangwb@marvell.com>