slab.h: ability to replace kmalloc with skmalloc
If you want, you can create new function named skmalloc to replace
kmalloc, this depend on the help of compiler tool.

User Guide:
If "ccflags-y += -DSUPPER_KMALLOC" is added in the Makefile of
driver or module, the kmalloc will be replace to be skmalloc.

So that you can do some advantage debug if need.

Change-Id: I1dff778bdfbc93d9fba54600709dddb39233d200
Signed-off-by: Jialing Fu <jlfu@marvell.com>
Signed-off-by: Lisa Du <cldu@marvell.com>