asoc: dapm: Add virtual single get/set function
Sometimes it is desirable to have a switch which does not reflect any
direct register configuration but which will instead only have an
effect implicitly. Provide a virtual single get/set function.

This new type is a virtual version of snd_soc_dapm_switch. It is used
when a backing register value is not necessary for deciding which
input path to connect.  A simple virtual value control e.g.
SOC_DAPM_SINGLE_VIRT() can be exposed to userspace which will be
used to choose which path to connect.

Change-Id: I91ee1683c78d9fada33addd789cbbe40349620d2
Signed-off-by: Nenghua Cao <nhcao@marvell.com>
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>