dma: mmp-tdma: add support for querying dma pointer position
currently the mmp tdma driver would update the prtd position by
adding the size of a period. This is accurate under the assuption
that the call back implemented accurately every period. but
sometimes under high load, it's impossible for the pos be updated
for an acurate period of time. so here using the actual DMA pointer
to update the position, and it reflacts the actual frame that DMA
currently moves when updating the position.

also change the dma burst size from 32 bytes to 4 bytes, which
impacts the real-time frame number reporting.

Change-Id: Ibe8daf964a004ecb1939862c6613b0d4b211285c
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>