ASoC: mmp-sspa: add audio mode ctrl
audio mode is used to set system voltage for audio scenarios. when
audio mode is enabled and with proper audio mode voltage is set,
low power mp3 in D1 will use the audio mode voltage, which is lower
and could save more power. also it doesn't require extra software to
control it.

Change-Id: I8dca7c07fc1b4188de98ca63bf56152e5bfe1fce
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>