char: 88pm830: change id gpadc initial sequence
according to silicon design team, for below sequence:
1. enable GPADC1_INT
2. set GPADC1_LOW_TH and GPADC1_UPP_TH
3. enable GPADC1_MEAS
GPADC1 interrupt will be triggered at step 2, which lead
to USB status change from undefined to a_idle. since USB
driver is not initialized when this interrupt is triggered.
in order to avoid triggering the GPADC1 interrupt when we
set GPADC1 threshold, we change the sequence to:
1. set GPADC1_LOW_TH and GPADC1_UPP_TH
2. enable GPADC1_INT
3. enable GPADC1_MEAS

Change-Id: I2157e01b6685f53b8c8f5c0f84fb1fdbd8766a1b
Signed-off-by: Fenghang Yin <yinfh@marvell.com>