clk: mmp: add DDR frequency change related support for ULC
From ULC, below modifications are added for DDR freq-chg part.
1. An extra pll1_1248 clock source is assigned to DDR to generate
624M PP.
1) Add pll1_1248 as ddr parent.
2) Share one DDR parent table on helan2 and ulc, ignore pll3p parent
as it will be only used for cpu.
3) Also add 624M PP for verification in case 667M DDR chip is used.
If 624M PP has real usage or not is determined by later test.

2. Add an extra LPM exit DDR DLL update table index configuration
in DFC register, 2 tables(0/1) could be used to identify if DLL
bypass mode is used or not. For bringup stage, always selects table0.
FIXME according to later requirement.

3. ULC supports max to 667M DDR, ddr_mode support is added to identify
if customer is using 533M/667M DDR. The major difference is PLL default
configuration and PP table.

Change-Id: I5b4ac57e6a9c3bee0fef25f44bdb346353359170
Signed-off-by: Zhoujie Wu <zjwu@marvell.com>