ASoC: soc-pcm: remove dpcm_path_put when found no path
remove dpcm_path_put when there is no path found in
dpcm_fe_dai_open, since it still uses this list when process path.

in map solution, we allow users to open audio card even when there
is not a completed audio path. so the logic here is unreasonable.
if not removing it, there will be kernel panic when trying to
activate a stream.

Change-Id: I9bfc6a60c695dfc27217c8bc57df04f2deb4fd06
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>