marvell: emmd: make emmd_page to be size alligned
The crashkernel size and offset is reserved by bootargs, like
crashkernel=4k@0x8140000

Now crashkernel is only used by the struct of emmd_page, but
it may be used by other purpose too in the furture.

In oder to avoid protetial confict, we reserve 4KB dedicatedly for
emmd_page, so let emmd_page to be 4KB aligned.

Change-Id: I84107923f50cc011c77691941ceff7695c357b5c
Signed-off-by: Jialing Fu <jlfu@marvell.com>