clk: mmp: add draft dvfs supprt for ulc
The major difference is from ulc, switch to use pmic
which support 8 VLs.

This is the draft version for ulc.

1) VL0 could be reserved for D1/D1P with lower cpu/ddr frequency,
avaliable for mp3 case.
2) VL1 is used for home screen.
3) VL2~3 could be used for video case.
4) VL4~6 could be used for gaming/camera/wfd cases.
5) VL7 for performance cases.

How to read fuse and frequency combination and svc table should
be adjusted after SVC spec is provided.

Note that ULC is using pm866, so pmic buck1 setting interface
will be adjusted later after pmic driver is ready.

Change-Id: Iaddc96bd2148f01375ed1235fab8d85cdfead6f9
Signed-off-by: Zhoujie Wu <zjwu@marvell.com>