ASoC: soc-pcm: revert pcm_buffer empty status checking patch
This reverts commit 7ee85fe3ec38feb6439f29f79586ae7efa666f25 on
kernel 3.10.

the previous patch fixes the issue that audio path close and DMA
interrupt contend, since audio path close will delete be_list,
while interrupt handler will access be_list.

Now there is another scenario that audio output is switched from
speaker to headset. Audio HAL will close BE while keeping audio
stream going on. It causes audio BE status not updated and speaker
(be) is not closed. This is a wrong status. speaker(be) is not used
and it's not closed.

According to this, we requested Audio HAL to ensure data transfer
ends before closing audio path. Then we don't need previous patch.

Already discussed with this issue in the community, and we need to
modify the current be_list update method and use more finer grained
spin lock to protect it.

Change-Id: I55d31e712a76bb179cce51fa00d6b27f95b6b711
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>