netfilter: have ip*t REJECT set the sock err when an icmp is to be sent
Author: JP Abgrall <jpa@google.com>
Date:   Thu Jun 16 14:37:03 2011 -0700

Allow the REJECT --reject-with icmp*blabla to also set the
matching error locally on the socket affected by the reject.
This allows the process to see an error almost as if it
received it via ICMP.
It avoids the local process who's ingress packet is rejected
to have to wait for a pseudo-eternity until some timeout kicks in.

Ideally, this should be enabled with a new iptables flag
similar to
 --reject-with-sock-err
For now it is enabled with CONFIG_IP*_NF_TARGET_REJECT_SKERR option.

Change-Id: I82cd819e13cccdb5e08092e877a63f3009e0853b
Signed-off-by: Gannicus Guo <jiaguo@marvell.com>