marvell: emmd: get emmd_page addr by phys_to_virt
Orignally, "pfn_to_page" and "page_address" were used to convert the
phys_addr to kaddr for emmd_page.

So no matter what the phys_addr is, the converted kaddr would be page
alligned.

But this is not a good idea:
1. In fact, struct emmd_page does not need to be page aligned.
2. If the phys_addr passed from bootloader is not page alligned, especially
when we switch to huge_page later; the kaddr we got may conflict with other
address.

So just use phys_to_virt instead.
Certainly, we still expect that bootloader can pass page aligned address
to kernel.

Change-Id: I43bb3e1343c69f8dcfdf7ac10b6d9322ef0664a8
Signed-off-by: Jialing Fu <jlfu@marvell.com>