ASoC: soc-pcm: fix dpcm_path_get list wrongly free
when dpcm_path_get returns with negative value, we should not call
dpcm_path_put to free list, since the list is not allocated actual
memory yet. Free an unknown pointer may result to unknown result,
or even kernel panic.

so once the return value is negative, we should exit the current
exec flow and return.

Change-Id: Icfc69b518bd1aa9ac75611f5fd973635d2c29011
Signed-off-by: Qiao Zhou <zhouqiao@marvell.com>