clocksource: timer-mmp: set IRQF_PERCPU to avoid migrate irq
When use SoC timer as local timer, need to set IRQF_PERCPU flag to avoid
migration of the timer irq. If without this flag, when doing hotplug,
the local timer irq may be migrated to other cores, and other cores can
execute the handler for current cpu, this is not allowed by locat timer.

Change-Id: I1b80bdf916a7d996259a55b156353bc2e6aab929
Signed-off-by: Leo Song <liangs@marvell.com>