arm64: dts: set soc index to 1 for pxa1928
For eden:
Android references a partition by the following path:
e.g. /dev/block/platform/soc.1/by-name/cache

Now in kernel 3.14, the soc index is 5.
But in kernel 3.10, the soc index is 1 (see the example above).

To make no change to android, we also set the soc index to 1
in kernel 3.14.

Change-Id: I622e958c57a52219e6b7381852fc586993d4fb1b
Signed-off-by: Xiang Wang <wangx@marvell.com>