ARM: dts: fix ground shifting issue for sp4555557sa battery
background:
after connected a charger, the battery start to charge and the
parasitic resistance between battery negative and board negative
pin causes a ground shifting which means the ground measured by
GPADC0 is not battery ground. for example, we assume Vp is the
voltage drop between battery negative and board negative, Vb is
the battery negative pad electric potential, Vg is the board
negative pad electric potential. obviously Vb = 0, but Vg = Vb +
Vp.
the ground shifting issue can affect battery temperature measurement.
for example, if we inject a bias current - Ibais, there is a voltage
drop between battery ground and GPADC0 - Vntc, but since there is
ground shifting, the NTC voltage measured by pm830 is Vntc - Vg.
so the ground shifting lead to a wrong battery temperature measurement.

how to fix it?
the resistance of the thermistor changes along with the temperature,
but the changes are not linear. for the sp4555557sa batteries, if we
inject 16uA bias current, when the temperature is below 15C,
V(ntc, t-1) - V(ntc, t) > 10mV, which means 10mV NTC voltage error can
lead to 1C battery temperature error. so if the ground shifting caused
30mV NTC voltage error, battery temperature error is less than 3C. but
if the temperature is above 15C, NTC voltage error will lead to a big
battery temperature error. so if the battery temperature is above 15C,
we need increase bias current to reduce the effecting of NTC voltage error.

solution:
to change bias current when the battery temperature crosses a certain
threshold (with 5C of hysteresis) by using GPADC0_INT. in this way,
because the bias current increases and the thermistor voltage increase
as well, the offset has less effect.
CASE1: -20C < Tbat < 18C
    - set Ibias = 16uA
    - set GPADC0_UPP_TH = LSB(-20C) and GPADC0_LOW_TH = LSB(18C)
CASE2: 12C < Tbat < 60C
    - set Ibias = 56uA
    - set GPADC0_UPP_TH = LSB(12C) and GPADC0_LOW_TH = LSB(60C)

Change-Id: I4cbd3063dab089e390d42b09c055f87256cbe72e
Signed-off-by: Fenghang Yin <yinfh@marvell.com>