video: mmp: add vdma clock dynamical control
STABILITY: there was one stability issue about vdma clock
control, it would cause system hang when vdma clock disabled unexpected
(when vdma was on). This patch was to control vdma clock in safe time.

POWER: there was some power contribution for vdma clock on even vdma
was disabled. So add vdma clock dynamical control.

Because of that vdma clock control was not shadowed, it could not
be disabled easily. The disable sequence should be:
>1. Disable VDMA channel and LCD object path.
  >a. Set bit 0 of VDMA_CH_CTRLx as 0.
  >b. Set bit 0 (video object) or 8 (graphic object) of
  LCD_PN_CTRL0/LCD_TV_CTRL0 as 0.
>2. Set both VDMA and LCD shadow flag.
>3. Disable target VDMA channel clock in VDMA_CLK_CTRL in
following display done interrupt handler (between display done
and VSYNC).

Change-Id: I5f9efa349ba293a7a4c625c36680efbeb7813538
Signed-off-by: Qing Zhu <qzhu@marvell.com>