vb2: fix bytesused == 0 handling
-in __verify_length() it would use the application-provided length
value for USERPTR and the vb2 core length for other memory models, but it
should have used the application-provided length as well for DMABUF.

-in __fill_vb2_buffer() on the other hand it would replace bytesused == 0
by the application-provided length, even for MMAP buffers where
the length is determined by the vb2 core.

-in __fill_vb2_buffer() it tries to figure out if all the planes have
bytesused == 0 before it will decide to replace bytesused by length.
However, the spec makes no such provision, and it makes for convoluted
code. So just replace any bytesused == 0 by the proper length.
The idea behind this was that you could use bytesused to signal empty
planes, something that is currently not supported. But that is better
done in the future by using one of the reserved fields in strucy v4l2_plane.

Change-Id: Ica7f93c4b19d794e0e33b9f4b9582b1b51c080aa
Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Owen Zhang <xinzha@marvell.com>